+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 - Fix some testes.
   [rafaelbco]
 
+- ported tests to plone.app.testing
+  [tomgross]
 
 1.5.10 (2014-04-16)
 -------------------
@@ -20,9 +22,6 @@ Changelog
 
 - Make sure mimetype is not None and use use filename for detection if available.
   [tschanzt]
-
-
-
 
 1.5.8 (2013-04-06)
 ------------------